Split main/worktree-ops.ts into worktree/{inspect,setup,remove,shared}.ts
